requirements.txt
google-adk
google-genai
python-dotenv
pydantic
sqlalchemy
aiosqlite
pyaudio
